+++ gitime/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'gitime'
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 __author__ = 'Jonathan Goodnow'
 __license__ = 'MIT License'
 __copyright__ = 'Copyright 2014 Jonathan Goodnow'