+++ gitime/database.py
@@ -3,12 +3,7 @@ import sqlite3
 import sys
 import os
 import stat
-
-# hack to get python 3 to not break when unicode is used
-try:
-    unicode
-except NameError:
-    unicode = str
+import six
 
 # locate the database
 DB_DIR = os.path.expanduser('~/.gitime')
@@ -31,21 +26,26 @@ def db_exists():
     return os.path.isfile(DB_NAME)
 
 
-def _db_connect(action):
+def _db_connect(action, new_db=False):
     """ Connects to the database, does something, and closes.
         Should only be called in `database.py`.
         Args:
         * action - a function with args 
                    (database connection object, database cursor)
+        Opts:
+        * new_db - if true, don't confirm that the database exists
     """
+    if not new_db and not db_exists():
+        print("Your database can't be found. Reinstalling should fix this.",
+            file=sys.stderr)
+        sys.exit()
     conn = results = None
     try:
         conn = sqlite3.connect(DB_NAME)
         c = conn.cursor()
         results = action(conn, c)
     except sqlite3.Error as e:
-        print("Database Error: %s" %e, file=sys.stderr)
-        sys.exit()
+        raise
     finally:
         if conn:
             conn.close()
@@ -85,7 +85,7 @@ def first_time_setup():
         """)
         conn.commit()
 
-    _db_connect(setup_action)
+    _db_connect(setup_action, True)
     if os.name in ('posix', 'mac'):
         set_unix_permissions(DB_NAME)
 
@@ -141,7 +141,7 @@ def query_invoice(unique):
             raise Exception("Rowid %d not in table invoice" %unique)
         return _query(lambda c: c.execute("SELECT * FROM invoice WHERE rowid=?", 
             (unique,)), False)
-    elif type(unique) in (str, unicode):
+    elif isinstance(unique, six.string_types):
         return _query(lambda c: c.execute("SELECT *, rowid FROM invoice WHERE name=?", 
             (unique,)), False)
     else: