+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ####dev
 
+* Using Six for Python 2 and 3 compatibility
+
+####1.0.1
+
 * Linux and Mac DB permissions correctly set
 * xlsxwriter now only required when exporting to xlsx
+* Better exception handling
+* Warns you if your database is missing
 
 ####1.0.0